requirements.txt
matplotlib==3.8.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
pyinstaller==6.6.0
PySide6==6.7.1
shapely==2.0.3
tqdm==4.66.2